Use POSIX qsort_r on FreeBSD >= 14
Prior to FreeBSD, qsort_r's arguments were the original BSD style,
where the thunk was the first argument to the comparison function.
However, the GNU style, where the thunk was the last argument, was
eventually adopted as the POSIX standard.  FreeBSD switched to this
beginning with FreeBSD 14.

So when compiling on FreeBSD, we first check to see whether to use the
BSD-style or GNU-style qsort_r declaration.

This patch is based on https://github.com/noporpoise/sort_r/pull/15